Added dark mode for project cards; Rearranged sections to gain a feel of things
