+++ src/components/LandingPage/index.tsx
@@ -9,8 +9,8 @@ import Projects from '~components/Projects'
 const LandingPage: FunctionComponent = () => {
   return (
     <div className={styles.container}>
-      <About />
       <Projects />
+      <About />
       <SectionLayout secondary title="Skills">
         test
       </SectionLayout>
