+++ src/svg/Site.tsx
@@ -11,11 +11,11 @@ const Site: React.FunctionComponent<Props> = ({ fillColor = 'white', width = '22
   <svg height={height} fill="none" viewBox="0 0 22 22" width={width} {...rest}>
     <path
       d="M14.2099 2.0625C14.2099 1.88016 14.2823 1.7053 14.4112 1.57636C14.5402 1.44743 14.715 1.375 14.8974 1.375H19.939C20.1214 1.375 20.2962 1.44743 20.4252 1.57636C20.5541 1.7053 20.6265 1.88016 20.6265 2.0625V7.10417C20.6265 7.2865 20.5541 7.46137 20.4252 7.5903C20.2962 7.71923 20.1214 7.79167 19.939 7.79167C19.7567 7.79167 19.5818 7.71923 19.4529 7.5903C19.324 7.46137 19.2515 7.2865 19.2515 7.10417V3.72167L13.5499 9.42333C13.4869 9.49088 13.411 9.54506 13.3267 9.58263C13.2424 9.62021 13.1513 9.64041 13.059 9.64204C12.9667 9.64367 12.875 9.62669 12.7894 9.59211C12.7038 9.55754 12.626 9.50607 12.5607 9.44079C12.4955 9.3755 12.444 9.29774 12.4094 9.21213C12.3748 9.12653 12.3579 9.03483 12.3595 8.94252C12.3611 8.85021 12.3813 8.75917 12.4189 8.67484C12.4565 8.59051 12.5106 8.51461 12.5782 8.45167L18.2799 2.75H14.8974C14.715 2.75 14.5402 2.67757 14.4112 2.54864C14.2823 2.4197 14.2099 2.24484 14.2099 2.0625Z"
-      fill="black"
+      fill={fillColor}
     />
     <path
       d="M2.29297 3.89591C2.29297 3.01041 3.01164 2.29175 3.89714 2.29175H11.6888C11.8711 2.29175 12.046 2.36418 12.1749 2.49311C12.3039 2.62204 12.3763 2.79691 12.3763 2.97925C12.3763 3.16158 12.3039 3.33645 12.1749 3.46538C12.046 3.59432 11.8711 3.66675 11.6888 3.66675H3.89714C3.83636 3.66675 3.77807 3.69089 3.73509 3.73387C3.69211 3.77685 3.66797 3.83514 3.66797 3.89591V18.1042C3.66797 18.2307 3.77064 18.3334 3.89714 18.3334H18.1055C18.1662 18.3334 18.2245 18.3093 18.2675 18.2663C18.3105 18.2233 18.3346 18.165 18.3346 18.1042V10.3126C18.3346 10.1302 18.4071 9.95538 18.536 9.82645C18.6649 9.69751 18.8398 9.62508 19.0221 9.62508C19.2045 9.62508 19.3793 9.69751 19.5083 9.82645C19.6372 9.95538 19.7096 10.1302 19.7096 10.3126V18.1042C19.7096 18.5297 19.5406 18.9377 19.2398 19.2386C18.9389 19.5394 18.5309 19.7084 18.1055 19.7084H3.89714C3.47168 19.7084 3.06366 19.5394 2.76282 19.2386C2.46198 18.9377 2.29297 18.5297 2.29297 18.1042V3.89591Z"
-      fill="black"
+      fill={fillColor}
     />
   </svg>
 )
